+++ elements/stacinfo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.3.0](https://github.com/EOX-A/EOxElements/compare/stacinfo-v0.2.1...stacinfo-v0.3.0) (2023-11-14)
+
+
+### Features
+
+* Allow rendering HTML in properties via config ([#419](https://github.com/EOX-A/EOxElements/issues/419)) ([76e10ed](https://github.com/EOX-A/EOxElements/commit/76e10edc2320348a8e80178a02225d90bea3c1af))
+
+## [0.2.1](https://github.com/EOX-A/EOxElements/compare/stacinfo-v0.2.0...stacinfo-v0.2.1) (2023-11-07)
+
+
+### Bug Fixes
+
+* Add styleOverride prop to stacinfo ([#390](https://github.com/EOX-A/EOxElements/issues/390)) ([7bff624](https://github.com/EOX-A/EOxElements/commit/7bff624b10530b609c7ebd2ef33711e8a6510b34))
+
+## [0.2.0](https://github.com/EOX-A/EOxElements/compare/stacinfo-v0.1.2...stacinfo-v0.2.0) (2023-10-24)
+
+
+### Features
+
+* Render single whitelisted property as full-width ([#366](https://github.com/EOX-A/EOxElements/issues/366)) ([bac0f23](https://github.com/EOX-A/EOxElements/commit/bac0f236ebf3f0eed2394673c9dafac69e4986de))
+
 ## 0.1.2 (2023-10-11)
 
 ### Miscellaneous Chores